refactor: move buffered message metrics to engine
This class is engine specific and should live next to the other engine
metrics, not next to the rocksdb layer.
