+++ engine/src/main/java/io/camunda/zeebe/engine/metrics/BufferedMessagesMetrics.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.db.impl.rocksdb;
+package io.camunda.zeebe.engine.metrics;
 
 import io.prometheus.client.Gauge;
